Refactoring; remove ununsed images, reorganized some files into better folder structure, renamed some files/classes/components.
