--- src/components/templates/Template.ts
@@ -1,5 +1,0 @@
-import { BaseProps } from "../BaseProps";
-
-export interface BaseTemplateProps extends BaseProps {
-    
-}
